fix: fully animate last modal out before closing the stack
This issue was happening because as soon as we were triggering the close of a modal, we were *also* animating the backdrop out, resulting in the last modal being "closed" before its out animation. Now, the backdrop will wait for that last modal stack item to be fully finished with its out animation, *before* checking if it should animate out itself.

Closes #96