fix: start animating out the last modal & the backdrop at the same time
The previous attempt at fixing this issue in 34879f was, unfortunately, not in the right direction.

The current approach aims at starting hiding both the last modal *and* the backdrop *at the same time*. That way, both animations can run concurrently. But! Only when the modal's animation will be done that the internal state will delete it from the stack's `openedItems` Set. Previously, we were updating the `openedItems` either before or during the animations, which was leading to these undesired behaviours.

However, it would be incumbent to the user to make sure that the animations config & timings provided to `createModalStack()` for both the modals and the backdrop work nicely together.

This is the another attempt at fixing this issue and hopefully the last one (🤞...).

Fixes #96
Closes #103