fix: prevent animations callback from being run only episodically
This issue was caused by a misuse of the `finished` value provided to `Animated.some_method().start()`'s callback: it simply indicates if an animation was run without interruption. It's it not a value that indicates if the animation is done being run and a callback can now be called.
As explain in the docs, simply providing a callback is actually enough, there was no need to check for `finished` truthiness; and in our case, that was actually causing bugs.

See: https://reactnative.dev/docs/animated#working-with-animations.

Fixes #120